requirements.txt
pymongo[srv]
flask
pymongo
bcrypt
pillow
gunicorn==19.6.0
request